implemented NodeDrawingContext for stateful styling: Node.RenderOpen(); Node.RenderOpenStatic() will return VisualDrawingContext for static or stateless. Changed: when trying to get a StyleRule from a StyleRuleSet, it will look for StyleRule for Normal state if specified style under certain state was not found.

+++ src/ImGui/Rendering/Pen.cs
@@ -6,13 +6,13 @@
     internal class Pen
     {
         public Color LineColor { get; set; } = Color.Black;
-        public float LineWidth { get; set; } = 1;
+        public double LineWidth { get; set; } = 1;
 
         public Pen()
         {
         }
 
-        public Pen(Color lineColor, float lineWidth)
+        public Pen(Color lineColor, double lineWidth)
         {
             LineColor = lineColor;
             LineWidth = lineWidth;
